Snowflake/etl_snowflake/requirements.txt
pyodbc 
snowflake-connector-python 
pandas
